requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
sumy==0.11.0
nltk==3.8.1
pyyaml==6.0.1
python-dotenv==1.0.0
lxml==4.9.3
# テスト用依存関係
pytest==7.4.3
pytest-mock==3.12.0